[add] いいねボタンをビューに追加、いいねの処理を厳密化

+++ app/controllers/favorites_controller.rb
@@ -1,15 +1,19 @@
 class FavoritesController < ApplicationController
   def create
     book = Book.find(params[:book_id])
-    favorite = current_user.favorites.new(book_id: book.id)
-    favorite.save
+    unless book.favorited_by?(current_user)
+      favorite = current_user.favorites.new(book_id: book.id)
+      favorite.save
+    end
     redirect_back(fallback_location: root_path)
   end
 
   def destroy
     book = Book.find(params[:book_id])
-    favorite = current_user.favorites.find_by(book_id: book.id)
-    favorite.destroy
+    if book.favorited_by?(current_user)
+      favorite = current_user.favorites.find_by(book_id: book.id)
+      favorite.destroy
+    end
     redirect_back(fallback_location: root_path)
   end
 end
